Prepare object indices for uint16_t (part 2)

+++ src/openrct2/object/ObjectManager.h
@@ -27,7 +27,7 @@ interface IObjectManager
     virtual Object* GetLoadedObject(size_t index) abstract;
     virtual Object* GetLoadedObject(int32_t objectType, size_t index) abstract;
     virtual Object* GetLoadedObject(const rct_object_entry* entry) abstract;
-    virtual uint8_t GetLoadedObjectEntryIndex(const Object* object) abstract;
+    virtual uint16_t GetLoadedObjectEntryIndex(const Object* object) abstract;
     virtual std::vector<rct_object_entry> GetInvalidObjects(const rct_object_entry* entries) abstract;
 
     virtual Object* LoadObject(const rct_object_entry* entry) abstract;
@@ -45,7 +45,7 @@ std::unique_ptr<IObjectManager> CreateObjectManager(IObjectRepository& objectRep
 
 void* object_manager_get_loaded_object_by_index(size_t index);
 void* object_manager_get_loaded_object(const rct_object_entry* entry);
-uint8_t object_manager_get_loaded_object_entry_index(const void* loadedObject);
+uint16_t object_manager_get_loaded_object_entry_index(const void* loadedObject);
 void* object_manager_load_object(const rct_object_entry* entry);
 void object_manager_unload_objects(const rct_object_entry* entries, size_t count);
 void object_manager_unload_all_objects();
